Remove FilterDataProvider class, only use FilterSettings class
