+++ src/main/java/de/uhd/ifi/se/decision/management/jira/model/impl/FilterSettingsImpl.java
@@ -3,8 +3,18 @@ package de.uhd.ifi.se.decision.management.jira.model.impl;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlRootElement;
+
 import org.codehaus.jackson.annotate.JsonProperty;
 
+import com.atlassian.jira.component.ComponentAccessor;
+import com.atlassian.jira.issue.issuetype.IssueType;
+import com.atlassian.jira.user.ApplicationUser;
+
+import de.uhd.ifi.se.decision.management.jira.filtering.QueryHandler;
 import de.uhd.ifi.se.decision.management.jira.model.DocumentationLocation;
 import de.uhd.ifi.se.decision.management.jira.model.FilterSettings;
 import de.uhd.ifi.se.decision.management.jira.model.KnowledgeType;
@@ -12,6 +22,8 @@ import de.uhd.ifi.se.decision.management.jira.model.KnowledgeType;
 /**
  * Model class for the filter settings.
  */
+@XmlRootElement(name = "issueTypesForDropdown")
+@XmlAccessorType(XmlAccessType.FIELD)
 public class FilterSettingsImpl implements FilterSettings {
 
 	private String projectKey;
@@ -20,6 +32,15 @@ public class FilterSettingsImpl implements FilterSettings {
 	private long createdLatest;
 	private List<DocumentationLocation> documentationLocations;
 	private List<KnowledgeType> issueTypes;
+	
+	@XmlElement
+	private List<String> allIssueTypes;
+	@XmlElement
+	private List<String> issueTypesMatchingFilter;
+	@XmlElement
+	private long startDate;
+	@XmlElement
+	private long endDate;
 
 	public FilterSettingsImpl() {
 		projectKey = "";
@@ -49,6 +70,44 @@ public class FilterSettingsImpl implements FilterSettings {
 		this.setDocumentationLocation(documentationLocations);
 		this.setIssueTypes(knowledgeTypes);
 	}
+	
+
+	public FilterSettingsImpl(String projectKey, String query, ApplicationUser user) {
+		this.allIssueTypes = new ArrayList<String>();
+		this.issueTypesMatchingFilter = new ArrayList<String>();
+		for (IssueType issueType : ComponentAccessor.getConstantsManager().getAllIssueTypeObjects()) {
+			this.allIssueTypes.add(issueType.getName());
+			this.issueTypesMatchingFilter.add(issueType.getName());
+		}
+
+		this.startDate = -1;
+		this.endDate = -1;
+
+		initFilterSettingsFromQuery(projectKey, query, user);
+
+		this.documentationLocations = new ArrayList<>();
+		DocumentationLocation[] locations = DocumentationLocation.values();
+		for (DocumentationLocation location : locations) {
+			this.documentationLocations.add(location);
+		}
+	}
+
+	public void initFilterSettingsFromQuery(String projectKey, String query, ApplicationUser user) {
+		if (!query.matches("\\?jql=(.)+") && !query.matches("\\?filter=(.)+")) {
+			return;
+		}
+		QueryHandler queryHandler = new QueryHandler(user, projectKey, false);
+
+		if (!queryHandler.getFilterSettings().getIssueTypes().isEmpty()) {
+			this.issueTypesMatchingFilter = new ArrayList<String>();
+			for (KnowledgeType type : queryHandler.getFilterSettings().getIssueTypes()) {
+				this.issueTypesMatchingFilter.add(type.toString());
+			}
+		}
+		this.startDate = queryHandler.getFilterSettings().getCreatedEarliest();
+		this.endDate = queryHandler.getFilterSettings().getCreatedLatest();
+	}
+
 
 	@Override
 	public String getProjectKey() {
@@ -104,6 +163,16 @@ public class FilterSettingsImpl implements FilterSettings {
 		}
 		return documentationLocations;
 	}
+	
+	@XmlElement(name="documentationLocations")
+	public List<String> getDocumentationLocations() {
+		List<String> documentationLocations = new ArrayList<String>();
+		DocumentationLocation[] locations = DocumentationLocation.values();
+		for (DocumentationLocation location : locations) {
+			documentationLocations.add(DocumentationLocation.getName(location));
+		}
+		return documentationLocations;
+	}
 
 	@Override
 	@JsonProperty("documentationLocationList")
@@ -137,5 +206,37 @@ public class FilterSettingsImpl implements FilterSettings {
 	@Override
 	public void setIssueTypes(List<KnowledgeType> types) {
 		issueTypes = types;
+	}
+	
+	public List<String> getAllIssueTypes() {
+		return allIssueTypes;
+	}
+
+	public void setAllIssueTypes(List<String> allIssueTypes) {
+		this.allIssueTypes = allIssueTypes;
+	}
+
+	public List<String> getIssueTypesMatchingFilter() {
+		return issueTypesMatchingFilter;
+	}
+
+	public void setIssueTypesMatchingFilter(List<String> issueTypesMatchingFilter) {
+		this.issueTypesMatchingFilter = issueTypesMatchingFilter;
+	}
+
+	public long getStartDate() {
+		return startDate;
+	}
+
+	public void setStartDate(long startDate) {
+		this.startDate = startDate;
+	}
+
+	public long getEndDate() {
+		return endDate;
+	}
+
+	public void setEndDate(long endDate) {
+		this.endDate = endDate;
 	}
 }
